refactor: move css-modules utilities out of index
