+++ src/index.js
@@ -22,11 +22,12 @@ import { isDir, isFile, stdout, isTruthy, removeScope } from './utils';
 import { getSizeInfo } from './lib/compressed-size';
 import { normalizeMinifyOptions } from './lib/terser';
 import {
-	parseMappingArgumentAlias,
+	parseAliasArgument,
 	parseMappingArgument,
 	toReplacementExpression,
 } from './lib/option-normalization';
 import { getConfigFromPkgJson, getName } from './lib/package-info';
+import { shouldCssModules, cssModulesConfig } from './lib/css-modules';
 
 // Extensions to use when resolving modules
 const EXTENSIONS = ['.ts', '.tsx', '.js', '.jsx', '.es6', '.es', '.mjs'];
@@ -309,9 +310,7 @@ function createConfig(options, entry, format, writeMeta) {
 		outputAliases['.'] = './' + basename(options.output);
 	}
 
-	const moduleAliases = options.alias
-		? parseMappingArgumentAlias(options.alias)
-		: [];
+	const moduleAliases = options.alias ? parseAliasArgument(options.alias) : [];
 	const aliasIds = moduleAliases.map(alias => alias.find);
 
 	const peerDeps = Object.keys(pkg.peerDependencies || {});
@@ -598,54 +597,4 @@ function createConfig(options, entry, format, writeMeta) {
 	};
 
 	return config;
-}
-
-function shouldCssModules(options) {
-	const passedInOption = processCssmodulesArgument(options);
-
-	// We should module when my-file.module.css or my-file.css
-	const moduleAllCss = passedInOption === true;
-
-	// We should module when my-file.module.css
-	const allowOnlySuffixModule = passedInOption === null;
-
-	return moduleAllCss || allowOnlySuffixModule;
-}
-
-function cssModulesConfig(options) {
-	const passedInOption = processCssmodulesArgument(options);
-	const isWatchMode = options.watch;
-	const hasPassedInScopeName = !(
-		typeof passedInOption === 'boolean' || passedInOption === null
-	);
-
-	if (shouldCssModules(options) || hasPassedInScopeName) {
-		let generateScopedName = isWatchMode
-			? '_[name]__[local]__[hash:base64:5]'
-			: '_[hash:base64:5]';
-
-		if (hasPassedInScopeName) {
-			generateScopedName = passedInOption; // would be the string from --css-modules "_[hash]".
-		}
-
-		return { generateScopedName };
-	}
-
-	return false;
-}
-
-/*
-This is done becuase if you use the cli default property, you get a primiatve "null" or "false",
-but when using the cli arguments, you always get back strings. This method aims at correcting those
-for both realms. So that both realms _convert_ into primatives.
-*/
-function processCssmodulesArgument(options) {
-	if (options['css-modules'] === 'true' || options['css-modules'] === true)
-		return true;
-	if (options['css-modules'] === 'false' || options['css-modules'] === false)
-		return false;
-	if (options['css-modules'] === 'null' || options['css-modules'] === null)
-		return null;
-
-	return options['css-modules'];
 }
